+++ R/gpkg-table.R
@@ -4,21 +4,6 @@
 gpkg_table_pragma <- function(x, table_name = NULL, ...)
   UseMethod("gpkg_table_pragma", x)
 
-#' @rdname gpkg_table
-#' @export
-gpkg_table_pragma.character <- function(x, table_name = NULL, ...) {
-  g <- geopackage(x, connect = TRUE)
-  res <- gpkg_table_pragma(g, table_name = table_name, ...)
-  gpkg_disconnect(g)
-  res
-}
-
-#' @rdname gpkg_table
-#' @export
-gpkg_table_pragma.SQLiteConnection <- function(x, table_name, ...) {
-  gpkg_table_pragma(geopackage(x), table_name, ...)
-}
-
 #' Lazy Access to Tables by Name
 #' 
 #' `gpkg_table_pragma()`: Get information on a table in a GeoPackage (without returning the whole table).
@@ -26,20 +11,23 @@ gpkg_table_pragma.SQLiteConnection <- function(x, table_name, ...) {
 #' @param x A _geopackage_ object or character path to GeoPackage file
 #' @param table_name _character_. One or more table names; for `gpkg_table_pragma()` if `table_name=NULL` returns a record for each table. `gpkg_table()` requires `table_name` be specified
 #' @param collect _logical_. Materialize a data.frame object in memory? Default: `FALSE` requires 'dbplyr' package. `TRUE` uses 'RSQLite'.
+#' @param column_names _character_. Used only when `collect=TRUE`. A _character_ vector of column names to select from `table_name`.
 #' @param query_string _logical_. Return SQLite query rather than executing it? Default: `FALSE`
 #' @param ... Additional arguments. In `gpkg_table()` arguments in `...` are passed to `dplyr::tbl()`. For `gpkg_table_pragma()`, `...` arguments are (currently) not used. For `gpkg_rast()` additional arguments are passed to `terra::rast()`. For `gpkg_vect()` additional arguments (such as `proxy=TRUE`) are passed to `terra::vect()`.
 #' @export
 #' @rdname gpkg_table
 #' @importFrom DBI dbGetQuery dbDisconnect
-gpkg_table_pragma.geopackage <- function(x, table_name = NULL, ...) {
+gpkg_table_pragma.default <- function(x, table_name = NULL, ...) {
   con <- .gpkg_connection_from_x(x)
-  tbls <- gpkg_list_tables(con)
-  dsn <- gpkg_source(x)
+  tbls <- gpkg_list_tables(x)
+  dsn <- con@dbname
+  
   if (is.null(table_name)) {
     table_name <- tbls
   }
   
-  if (!all(table_name %in% tbls)) stop("no table with name: '", paste0(table_name[!table_name %in% tbls], collapse = "', '"), "' in ", dsn)
+  if (!all(table_name %in% tbls)) 
+    stop("no table with name: '", paste0(table_name[!table_name %in% tbls], collapse = "', '"), "' in ", dsn)
   
   res <- do.call('rbind', lapply(table_name, function(xx) {
     data.frame(dsn = dsn, 
@@ -59,10 +47,9 @@ gpkg_table_pragma.geopackage <- function(x, table_name = NULL, ...) {
 
 #' @export
 #' @rdname gpkg_table
-#' @examplesIf !inherits(try(requireNamespace("RSQLite", quietly = TRUE)), 'try-error') &&!inherits(try(requireNamespace("dbplyr", quietly = TRUE)), 'try-error') && !inherits(try(requireNamespace("terra", quietly = TRUE)), 'try-error')
-#' @description `gpkg_table()`: access a specific table (by name) and get a "lazy" `tibble` object referencing that table
+#' @description `gpkg_table()`: Access a specific table (by name) and get a _tbl_SQLiteConnection_ object referencing that table
 #' @return `gpkg_table()`: A 'dbplyr' object of class _tbl_SQLiteConnection_
-#' @examples 
+#' @examplesIf requireNamespace("RSQLite", quietly = TRUE) && requireNamespace("dbplyr", quietly = TRUE) && requireNamespace("terra", quietly = TRUE)
 #' 
 #' tf <- tempfile(fileext = ".gpkg")
 #' 
@@ -79,12 +66,12 @@ gpkg_table_pragma.geopackage <- function(x, table_name = NULL, ...) {
 #'            RASTER_TABLE = "DEM2",
 #'            FIELD_NAME = "Elevation")
 #'
-#' g <- geopackage(tf)
+#' g <- geopackage(tf, connect = TRUE)
 #' 
 #' # inspect gpkg_contents table
 #' gpkg_table(g, "gpkg_contents")
 #' 
-#' gpkg_vect(g, "gpkg_contents")
+#' gpkg_contents(g)
 #' 
 #' # materialize a data.frame from gpkg_2d_gridded_tile_ancillary
 #' library(dplyr, warn.conflicts = FALSE)
@@ -98,6 +85,7 @@ gpkg_table_pragma.geopackage <- function(x, table_name = NULL, ...) {
 gpkg_table <- function(x,
                        table_name,
                        collect = FALSE,
+                       column_names = "*",
                        query_string = FALSE,
                        ...)
   UseMethod("gpkg_table", x)
@@ -107,43 +95,71 @@ gpkg_table <- function(x,
 gpkg_table.default <- function(x,
                                table_name,
                                collect = FALSE,
+                               column_names = "*",
                                query_string = FALSE,
                                ...) {
-    
+  
+  if (is.null(column_names) ||
+      length(column_names) == 0 ||
+      nchar(as.character(column_names)) == 0) {
+    column_names <- "*"
+  }
+  
+  q <- sprintf("SELECT %s FROM %s",
+               paste0(column_names, collapse = ", "),
+               table_name)
+  
+  if (isTRUE(query_string)) {
+    return(q)
+  }
+  
   con <- .gpkg_connection_from_x(x)
   
-  if (isTRUE(collect) || isTRUE(query_string)) {
-    
+  if (isTRUE(collect)) { 
+    res <- gpkg_query(con, q)
     if (attr(con, 'disconnect')) {
-      on.exit(DBI::dbDisconnect(con))
+      gpkg_disconnect(con)
     }
-    
-    q <- sprintf("SELECT * FROM %s", table_name)
-    if (query_string) {
-      return(q)
-    }
-    
-    return(gpkg_query(con, q))
+    return(res)
   }
   
   stopifnot(requireNamespace("dbplyr", quietly = TRUE))
   
-  tbls <- gpkg_list_tables(con)
+  res <- try(dplyr::tbl(con, table_name, ...), silent = FALSE)
   
-  if (missing(table_name) || length(table_name) == 0) stop("table name should be one of:", paste0(tbls, collapse = ", "), call = FALSE)
+  if (inherits(res, 'try-error')) {
+    tbls <- gpkg_list_tables(con)
+
+    if (length(tbls) == 0) {
+      tbls <- "<none available>"
+    }
+    
+    stop("table name should be one of: ",
+         paste0(tbls, collapse = ", "), call. = FALSE)
+  }
   
-  dplyr::tbl(con, table_name, ...)
+  # keep tbl_SQLiteconnection open
+  if (attr(con, 'disconnect')) {
+    attr(con, 'disconnect') <-  FALSE
+  }
+  
+  res
 }
 
-#' @description `gpkg_collect()`: alias for `gpkg_table(..., collect=TRUE)`
+#' @description `gpkg_collect()`: Alias for `gpkg_table(..., collect=TRUE)`
 #' @return `gpkg_collect()`: An object of class _data.frame_
 #' @rdname gpkg_table
 #' @export
 gpkg_collect <- function(x, table_name, query_string = FALSE, ...) {
+  
+  if (!requireNamespace("RSQLite", quietly = TRUE)) {
+    stop('package `RSQLite` is required to get the `gpkg_extensions` table', call. = FALSE)
+  }
+  
   gpkg_table(x, table_name, ..., query_string = query_string, collect = TRUE)
 }
 
-#' @description `gpkg_tbl()`: shorthand for `gpkg_table(..., collect=FALSE)`(default) that always returns a 'dplyr' object.
+#' @description `gpkg_tbl()`: Alias for `gpkg_table(..., collect=FALSE)`(default) that _always_ returns a _tbl_SQLiteConnection_ object.
 #' @return `gpkg_tbl()`: An object of class _tbl_SQLiteConnection_
 #' @rdname gpkg_table
 #' @export
@@ -151,6 +167,7 @@ gpkg_tbl <- function(x, table_name, ...) {
   gpkg_table(x, table_name, ..., collect = FALSE)
 }
 
+#' @description `gpkg_rast()`: Get a _SpatRaster_ object corresponding to the specified `table_name`
 #' @return `gpkg_rast()`: A 'terra' object of class _SpatRaster_
 #' @export
 #' @rdname gpkg_table
@@ -167,6 +184,7 @@ gpkg_rast <- function(x, table_name = NULL, ...) {
 }
 
 
+#' @description `gpkg_vect()`: Get a _SpatVector_ object corresponding to the specified `table_name`
 #' @return `gpkg_vect()`: A 'terra' object of class _SpatVector_ (may not contain geometry columns)
 #' @export
 #' @rdname gpkg_table
@@ -176,12 +194,45 @@ gpkg_vect <- function(x, table_name, ...) {
   res <- try(terra::vect(x$dsn, layer = table_name, ...), silent = TRUE)
   if (inherits(res, 'try-error')) {
     # create features, try again with layer not specified
-    gpkg_create_dummy_features(x)
+    # gpkg_create_empty_features(x, table_name = "dummy_features", contents = FALSE)
     res2 <- try(terra::vect(x$dsn, query = paste("SELECT * FROM", table_name), ...), silent = TRUE)
     if (inherits(res2, 'try-error')) {
       stop(res2[1], call. = FALSE)
     }
     res <- res2
   }
+  res
+}
+
+#' @description `gpkg_sf()`: Get a _sf-tibble_ object corresponding to the specified `table_name`
+#' @return `gpkg_sf())`: An _sf-tibble_ object of class `"sf"`, `"tbl_df"`. If the table contains no geometry column the result is a `"tbl_df"`.
+#' @export
+#' @rdname gpkg_table
+gpkg_sf <- function(x, table_name, ...) { 
+  if (!requireNamespace("sf", quietly = TRUE))
+    stop("package 'sf' is required to create 'sf' data.frame from tables in a GeoPackage", call. = FALSE)
+  x <- .gpkg_connection_from_x(x)
+  try(sf::read_sf(x$dsn, layer = table_name, ...), silent = TRUE)
+}
+
+gpkg_create_table <- function(x, table_name, fields, ...) {
+  con <- .gpkg_connection_from_x(x)
+  res <- try(RSQLite::dbCreateTable(con, name = table_name, fields = fields, ...))
+  
+  if (attr(con, 'disconnect')) {
+    gpkg_disconnect(con) 
+  }
+  
+  res
+}
+
+gpkg_append_table <- function(x, table_name, value, ...) {
+  con <- .gpkg_connection_from_x(x)
+  res <- try(RSQLite::dbAppendTable(con, name = table_name, value = value, ...))
+  
+  if (attr(con, 'disconnect')) {
+    gpkg_disconnect(con) 
+  }
+  
   res
 }